requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
python-telegram-bot==20.7
python-dotenv==1.0.0
yt-dlp
